+++ CH341WindllDriver.py
@@ -1,0 +1,3 @@
+from ctypes import *
+
+driver = windll.LoadLibrary("CH341DLL.dll")
